+++ lib/smart_form_fields.dart
@@ -40,7 +40,8 @@ export 'src/fields/smart_email_field.dart' show SmartEmailField;
 export 'src/fields/smart_form_field.dart'
     show SmartFieldBuilder, SmartFormField;
 export 'src/fields/smart_password_field.dart' show SmartPasswordField;
-export 'src/fields/smart_phone_field.dart' show SmartPhoneField;
+export 'src/fields/smart_phone_field.dart'
+    show SmartPhoneCountrySelectorLayout, SmartPhoneField;
 export 'src/fields/smart_text_field.dart' show SmartTextField;
 export 'src/validation/smart_async_validator.dart' show SmartAsyncValidator;
 export 'src/validation/smart_async_validators.dart' show SmartAsyncValidators;